% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.ec2metadata-package.R
\docType{package}
\name{aws.ec2metadata-package}
\alias{aws.ec2metadata}
\alias{aws.ec2metadata-package}
\alias{metadata}
\title{Get EC2 Instance Metadata}
\format{An object of class \code{list} of length 24.}
\usage{
metadata
}
\description{
Retrieve EC2 instance metadata from the instance
}
\details{
This is a list of functions to return various metadata values for the currently running EC2 instance. These are only meant to be called from an EC2 instance; no guarantees are made about behavior on other platforms. Two functions: \code{versions()} and \code{meta_data_items()} return the versions of the metadata, and the set of top-level metadata items, respectively.

The function \code{item()} retrieves a particular metadata item specified by its full path.

The remaining functions in the list are aliases for potentially commonly needed metadata items

Generally, this will return a character string containing the requested information, otherwise a \code{NULL} if the response is empty. The \code{iam_role()} function returns a list. An error will occur if, for some reason, the request otherwise fails.
}
\examples{
names(metadata)

\dontrun{
  metadata$versions()
  metadata$items()

  # get instance id
  metadata$instance_id()
  # get ami id
  metadata$ami_id()
  
  # get IAM role (NULL if none specified)
  metadata$iam_info()
  metadata$iam_role("myrole")

  # get an arbitrary metadata item
  metadata$item("meta-data/placement/availability-zone")
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-metadata.html}{Metadata Documentation}
}
\keyword{datasets}
\keyword{package}

